Upd: refactor controllers
