+++ server/serverAPI/controller/BaseController.ts
@@ -1,0 +1,9 @@
+export default class BaseController {
+    protected getException(error: unknown): string {
+        if (error instanceof Error) {
+            return error.message;
+        }
+
+        return String(error);
+    }
+}
