+++ server/serverAPI/controller/UserController.ts
@@ -4,95 +4,16 @@ import ResponseFormatter from "../../utils/ResponseFormatter";
 import { ResponseTypes } from "../../utils/ResponseTypes";
 import IBaseUser from "../model/user/IBaseUser";
 import IUser from "../model/user/IUser";
-import UserUpdateSchema from "../model/user/requestSchema/UserUpdateSchema";
+import BaseUserController from "./BaseUserController";
 
 /**
  * This class creates several properties responsible for user-actions 
  * provided to the user.
  */
-export default class UserController {
-    private database: IDatabase<IUser>;
-
+export default class UserController extends BaseUserController {
     constructor(database: IDatabase<IUser>) {
-        this.database = database;
+        super(database);
     }
-
-    private getException(error: unknown): string {
-        if (error instanceof Error) {
-            return error.message;
-        }
-
-        return String(error);
-    }
-
-    private convertToBaseUser(user: IUser): IBaseUser {
-        return {
-            firstName: user.firstName,
-            lastName: user.lastName,
-            lastSeen: user.lastSeen
-        };
-    }
-
-    private async validateUser(req: Request, res: Response, user: IUser): Promise<IUser> {
-        let userSchema = new UserUpdateSchema(
-            req.body.firstName === undefined ? user.firstName : req.body.firstName,
-            req.body.lastName === undefined ? user.lastName : req.body.lastName,
-            req.body.username === undefined ? user.username : req.body.username,
-            req.body.password === undefined ? user.password : req.body.password,
-            user.lastSeen,
-        );
-
-        let logs = await userSchema.validate();
-
-        if (logs.length > 0) {
-            return Promise.reject(res.status(400)
-                .json(ResponseFormatter.formatAsJSON(ResponseTypes.ERROR, logs)));
-        }
-
-        let newUser: IUser = {
-            inventory: user.inventory,
-            firstName: user.firstName,
-            lastName: user.lastName,
-            lastSeen: user.lastSeen,
-            password: user.password,
-            username: user.username
-        };
-
-        return Promise.resolve(newUser);
-    }
-
-    private async updateUser(req: Request, res: Response, user: IUser): Promise<IUser> {
-        return this.database.Update(req.serverUser.username, user).then(updatedUser => {
-            if (updatedUser === null) {
-                return Promise.reject(res.status(400)
-                    .json(ResponseFormatter.formatAsJSON(ResponseTypes.ERROR, "User update error.")));
-            }
-
-            return Promise.resolve(updatedUser);
-        }, (error) => {
-            return Promise.reject(res.status(400)
-                .json(ResponseFormatter.formatAsJSON(ResponseTypes.ERROR, this.getException(error))));
-        });
-    }
-
-    private async getUser(req: Request, res: Response): Promise<IUser> {
-        let parameters = new Map<string, any>([
-            ["username", req.serverUser.username]
-        ]);
-
-        return this.database.Get(parameters).then(async user => {
-            if (user === null) {
-                return Promise.reject(res.status(404)
-                    .json(ResponseFormatter.formatAsJSON(ResponseTypes.ERROR, "User hasn't been found")));
-            }
-
-            return Promise.resolve(user);
-        }, (error) => {
-            return Promise.reject(res.status(400)
-                .json(ResponseFormatter.formatAsJSON(ResponseTypes.ERROR, this.getException(error))));
-        });
-    }
-
 
     /**
      * Lets client to get information about all users existed on the server.
