+++ server/serverAPI/model/user/IContactInformation.ts
@@ -1,0 +1,3 @@
+export default interface IContactInformation {
+    email?: string
+}
