+++ server/spoonacularUtils/SpoonacularAPI.ts
@@ -1,4 +1,4 @@
-import axios, { AxiosResponse } from "axios";
+import axios from "axios";
 import RequestLimitReached from "../exceptions/RequestLimitReached";
 
 export default abstract class SpoonacularAPI {
